Reorganize profile code
